feat: optimize chart for easy visiblity
